Put experimental tests with code instead of in operator
