+++ dask_kubernetes/conftest.py
@@ -1,13 +1,24 @@
 import pytest
 
+from glob import glob
+import pathlib
 import os
 import subprocess
 
+from kopf.testing import KopfRunner
+
 from dask_kubernetes.common.utils import check_dependency
+
+DIR = pathlib.Path(__file__).parent.absolute()
 
 check_dependency("helm")
 check_dependency("kubectl")
 check_dependency("docker")
+
+
+@pytest.fixture()
+async def kopf_runner(k8s_cluster):
+    yield KopfRunner(["run", "-m", "dask_kubernetes.operator", "--verbose"])
 
 
 @pytest.fixture(scope="session")
@@ -28,3 +39,11 @@ def k8s_cluster(kind_cluster, docker_image):
 @pytest.fixture(scope="session")
 def ns(k8s_cluster):
     return "default"
+
+
+@pytest.fixture(scope="session", autouse=True)
+def customresources(k8s_cluster):
+    crd_path = glob(os.path.join(DIR, "operator", "customresources"))
+    k8s_cluster.kubectl("apply", "-f", *crd_path)
+    yield
+    k8s_cluster.kubectl("delete", "-f", *crd_path)
